requirements.txt
# VulnHawk - Network and Host Vulnerability Scanner
# Required Dependencies

python-nmap>=0.7.1
requests>=2.28.0
rich>=13.0.0
jinja2>=3.1.0
flask>=2.3.0
python-dotenv>=1.0.0
